fix(non-sec): F7 extract record_files test fixture builders to store/tests/
- Move all 7 test helper/fixture builders (causal, user_book_payload,
  user_book_count, vocab_payload, fnv1a_name, write_legacy_fnv_record,
  single_pref_record) out of the 4255L god module into
  store/tests/record_files_helpers.rs; record_files.rs 4255 -> 4176L
- #[cfg(test)] #[path]-included child module keeps private access to
  record_files items (no visibility bumps, store/mod.rs untouched)
- F15 verified already complete: no CSP in index.html (src/built) or
  tauri.conf.json since 93a1828; only response.rs runtime CSP remains
- Verified: cargo check + cargo check --tests green, 53 record_files
  tests pass, check:frontend green
